+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 - Easily authenticate to a [ECR](https://aws.amazon.com/ecr/) registry: `awless authenticate registry`
 - Listing [ECS](https://aws.amazon.com/ecs/) clusters, services and containers: `awless list containerclusters/containerservices/containers`
 - Create/Delete [ECS](https://aws.amazon.com/ecs/) cluster or container: `awless create/delete containercluster/container`
-- Start/Stop container services: `awless start/stop containerservice`
+- Start/Stop container services or tasks: `awless start/stop containerservice`, `awless start containertasks`
 - Table display now use full terminal width when possible